+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 
 ### Changed
 
+- Improved UI by hiding exercise action buttons (Save/Delete) by default and showing them only on hover or when clicking the "More" menu
+- Added "More" menu button (⋮) for better touch device compatibility when accessing exercise actions
+
 ### Fixed
 
 - Fixed issue where the collapsed/expanded state of training units was lost when adding new units